+++ scripts/lichessTools-init.js
@@ -104,7 +104,6 @@ lichessTools.loadTool(LiChessTools.Tools.CommonTeams);
 lichessTools.loadTool(LiChessTools.Tools.ExpandAll);
 lichessTools.loadTool(LiChessTools.Tools.StudyLinks);
 lichessTools.loadTool(LiChessTools.Tools.Konami);
-lichessTools.loadTool(LiChessTools.Tools.AnalysisWheel);
 lichessTools.loadTool(LiChessTools.Tools.CevalLineOptions);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);