+++ scripts/lichessTools-start.js
@@ -1,1 +1,3 @@
-lichessTools.start(window.site);
+const start = ()=>setTimeout(()=>lichessTools.start(window.site));
+const load = window.site?.load;
+load?.then(start);