+++ tools/commonTeams/tool.js
@@ -38,16 +38,6 @@
       if (!crosstable.length || !parent.inViewport(crosstable)) return;
       const commonTeamsLink=$('a.lichessTools-commonTeams');
       if (commonTeamsLink.length) return;
-      const userId=parent.getUserId();
-      const isMyGame = !!$('.game__meta__players a.user-link')
-        .get()
-        .find(e=>{
-          const href=$(e).attr('href');
-          if (!href) return;
-          const isPlayer=href.toLowerCase().includes(userId.toLowerCase());
-          return isPlayer;
-        });
-      //if (!isMyGame) return;
       const teamsArr=[];
       const promises=[];
       $('.game__meta__players a.user-link')