+++ tools/friendsList/tool.js
@@ -9,14 +9,16 @@
         category: 'friends',
         type:'single',
         possibleValues: ['default','open','menu','button','hidden'],
-        defaultValue: 'menu'
+        defaultValue: 'menu',
+        needsLogin: true
       },
       {
         name:'liveFriendsPage',
         category: 'friends',
         type:'single',
         possibleValues: [false,true],
-        defaultValue: true
+        defaultValue: true,
+        needsLogin: true
       }
     ];
 
@@ -620,8 +622,10 @@
       };
       const lichess=parent.lichess;
       if (!lichess) return;
-      const userId=lichessTools.getUserId();
-      if (!userId) return;
+      if (!parent.getUserId()) {
+        parent.global.console.debug(' ... Disabled (not logged in)');
+        return;
+      }
       const setInterval=parent.global.setInterval;
       const clearInterval=parent.global.clearInterval;
       lichess.pubsub.off('socket.in.following_onlines', this.following_onlines);
@@ -660,7 +664,7 @@
             clearInterval(this.onlinesInterval);
             this.getFollowingOnlinesByApi();
           }
-        },1000);
+        },5000);
       }
 
       switch(friendsBoxMode) {