+++ tools/preferences/tool.js
@@ -28,6 +28,7 @@
         'options.advancedPreferences': 'Advanced preferences',
         'author': 'by %s',
         'lichessTools': 'LiChess Tools',
+        'lichessToolsPreferences': 'LiChess Tools preferences',
         'feedbackButtonTitle': 'Send feedback about LiChess Tools',
         'feedbackTitle': 'Send a message to the developer',
         'resetButtonText': 'Reset',
@@ -54,6 +55,7 @@
         'options.advancedPreferences': 'Preferin\u0163e avansate',
         'author': 'de %s',
         'lichessTools': 'LiChess Tools',
+        'lichessToolsPreferences': 'Preferin\u0163e LiChess Tools',
         'feedbackButtonTitle': 'Trimite p\u0103reri despre LiChess Tools',
         'feedbackTitle': 'Trimite un mesaj programatorului',
         'resetButtonText': 'Resetare',
@@ -81,11 +83,9 @@
     const applyOptions=parent.applyOptions;
     const lichess=parent.lichess;
     const isOptionSet=parent.isOptionSet;
+    const isLoggedIn=!!parent.getUserId();
 
-    const lichessToolsText=trans.noarg('lichessTools');
-    if (parent.global.document.title.indexOf(lichessToolsText)<0) {
-      parent.global.document.title=parent.global.document.title?.replace(/^[^\s]+/,lichessToolsText);
-    }
+    parent.global.document.title=trans.noarg('lichessToolsPreferences');
 
     $('nav.page-menu__menu a.active').removeClass('active');
     $('a.lichessTools-menu').addClass('active');
@@ -134,17 +134,23 @@
         </tr>
     </tbody>
 </table>
-<h3>$trans(feedbackTitle)</h3>
+`;
+    if (isLoggedIn) {
+      html+=`<h3>$trans(feedbackTitle)</h3>
 <div class="feedback">
   <textarea enterkeyhint="send"></textarea>
   <button data-icon="&#xE03A;" title="$trans(feedbackButtonTitle)"></button>
 </div>`;
+    }
         
     const categs={};
     for (const tool of tools) {
       if (!tool.preferences) continue;
       for (const pref of tool.preferences) {
-        if (pref.hidden && !parent.debug) continue;
+        if (!parent.debug) {
+          if (pref.hidden) continue;
+          if (!isLoggedIn && pref.needsLogin) continue;
+        }
         let categ=categs[pref.category];
         if (!categ) {
           categ=[];
@@ -158,10 +164,12 @@
       const categ=categs[key];
       html+='<div><h3><label for="chk_'+key+'">$trans(options.'+key+')</label></h3><input type="checkbox" id="chk_'+key+'" class="categoryToggle">';
       for (const pref of categ) {
+        const defaultValue=(!isLoggedIn && pref.defaultNotLoggedInValue!==undefined) ? pref.defaultNotLoggedInValue : pref.defaultValue;
+
         html+=`<section data-pref="${pref.name}"`;
         const classes=[];
         if (pref.advanced) classes.push('lichessTools-advancedPreference');
-        if (pref.hidden) classes.push('lichessTools-hiddenPreference');
+        if (pref.hidden||(!isLoggedIn&&pref.needsLogin)) classes.push('lichessTools-hiddenPreference');
         if (pref.wip) classes.push('lichessTools-wipPreference');
         if (classes.length) html+=' class="'+classes.join(' ')+'"'
         html+=`><h2>$trans(options.${pref.name})`;
@@ -176,7 +184,7 @@
               const textKey=typeof val==='boolean'
                 ? (val?'yes':'no')
                 : (pref.valuePrefix||pref.name+'.')+val;
-              html+=`<div`+(parent.isOptionSet(pref.defaultValue,val)?' class="defaultValue"':'')+`>
+              html+=`<div`+(parent.isOptionSet(defaultValue,val)?' class="defaultValue"':'')+`>
                   <input type="radio" value="${val}" name="${pref.name}"/>
                   <label>$trans(${textKey})</label>
                 </div>`;
@@ -190,7 +198,7 @@
               const textKey=typeof val==='boolean'
                 ? (val?'yes':'no')
                 : (pref.valuePrefix||pref.name+'.')+val;
-              html+=`<div`+(parent.isOptionSet(pref.defaultValue,val)?' class="defaultValue"':'')+`>
+              html+=`<div`+(parent.isOptionSet(defaultValue,val)?' class="defaultValue"':'')+`>
                   <input type="checkbox" value="${val}" name="${pref.name}"/>
                   <label>$trans(${textKey})</label>
                 </div>`;
@@ -252,7 +260,8 @@
           .find(p=>p.name==optionName);
         if (currentValue!==undefined) {
           if (isCheckable) {
-            const checked = isOptionSet(currentValue,optionValue,preferences?.defaultValue);
+            const defaultValue=(!isLoggedIn && preferences?.defaultNotLoggedInValue!==undefined) ? preferences?.defaultNotLoggedInValue : preferences?.defaultValue;
+            const checked = isOptionSet(currentValue,optionValue,defaultValue);
             $(e).prop('checked',checked);
           }
           else {
@@ -306,7 +315,14 @@
           ev.preventDefault();
           if (!parent.global.confirm(trans.noarg('resetButtonWarning'))) return;
           const options=await parent.getOptions();
-          const data=parent.tools.map(t=>t.preferences).flat().filter(p=>p).map(p=>({name:p.name,value:p.defaultValue}));
+          const data=parent.tools
+                       .map(t=>t.preferences)
+                       .flat()
+                       .filter(p=>p)
+                       .map(p=>{
+                         const defaultValue=(!isLoggedIn && p.defaultNotLoggedInValue!==undefined) ? p.defaultNotLoggedInValue : p.defaultValue;
+                         return { name:p.name,value:defaultValue };
+                       });
           for (const {name,value} of data) options[name]=value;
           await applyOptions(options);
           parent.fireReloadOptions();
@@ -366,13 +382,6 @@
       checkGlobalSwitch();
       checkAdvanced();
       this.addInfo();
-      const m=/#lichessTools\/(?<pref>.*)$/.exec(parent.global.location.hash);
-      const pref=m?.groups?.pref;
-      if (pref) {
-        const elem=$('[data-pref="'+pref+'"]');
-        elem.parents().add(elem).show();
-        elem[0]?.scrollIntoView();
-      }
     };
 
     addInfo() {
@@ -390,9 +399,31 @@
       });
     };
 
+    addPreferencesMenu=(isOpening)=>{
+      if (!isOpening) return;
+      const parent=this.lichessTools;
+      const trans=parent.translator;
+      const $=parent.$;
+      if (!$('#dasher_app a.lichessTools-preferences').length) {
+        let links=$('#dasher_app div.links');
+        if (!links.length) {
+          links=$('<div class="links">')
+            .insertBefore('#dasher_app div.subs');
+        }
+        $('<a class="text lichessTools-preferences">')
+          .attr('data-icon','\uE019')
+          .attr('href','/team/all#lichessTools')
+          .text(trans.noarg('lichessTools'))
+          .attr('title',trans.noarg('lichessToolsPreferences'))
+          .appendTo(links);
+      }
+
+    };
+
     async start() {
       const parent=this.lichessTools;
       const $=parent.$;
+      const lichess=parent.lichess;
       const location=parent.global.location;
       const trans=parent.translator;
       this.options={
@@ -401,13 +432,39 @@
       };
       this.logOption('Integration in Preferences', this.options.enabled);
       this.logOption(' ... show advanced', this.options.advanced);
-      if (!$('main.account').length) return;
+
+      const isLoggedIn=!!parent.getUserId();
+      const isLoggedOutTeams=location.pathname=='/team/all' && !isLoggedIn;
+      lichess.pubsub.off('dasher.toggle',this.addPreferencesMenu);
+      if (!isLoggedIn) {
+        lichess.pubsub.on('dasher.toggle',this.addPreferencesMenu);
+      }
+
+      if (!$('main.account').length && !isLoggedOutTeams) {
+        return;
+      }
+
       if ($('a.lichessTools-menu').length) return;
       const openPreferences=this.openPreferences;
 
+      let $this=this;
       const f=function() {
         if (location.hash?.startsWith('#lichessTools')) {
+          if (isLoggedOutTeams) {
+            lichess.asset.loadCssPath('bits.account');
+            $('main nav.subnav').hide();
+          }
           openPreferences();
+          parent.global.setTimeout(()=>{
+            const m=/#lichessTools\/(?<pref>.*)$/.exec(location.hash);
+            const pref=m?.groups?.pref;
+            if (pref!=$this._lastScrolled) {
+              const elem=$('[data-pref="'+pref+'"]');
+              elem.parents().add(elem).show();
+              elem[0]?.scrollIntoView();
+              $this._lastScrolled=pref;
+            }
+          },500);
         } else {
           if ($('.lichessTools-preferences').length) {
             location.reload();
@@ -418,7 +475,7 @@
 
       const prefElem=$('<a>')
         .addClass('lichessTools-menu')
-        .attr('href','/account/preferences/display#lichessTools')
+        .attr('href',isLoggedOutTeams?'/team/all#lichessTools':'/account/preferences/display#lichessTools')
 	    .append($('<span>').text(trans.noarg('LiChess Tools')));
 
       $('.page-menu__menu a[href*=privacy]').before(prefElem);