+++ tools/timelineNotify/tool.js
@@ -26,7 +26,8 @@
           'plan-renew'*/
         ],
         defaultValue: 'forum-post,ublog-post',
-        advanced: true
+        advanced: true,
+        needsLogin: true
       }
     ];
 
@@ -97,7 +98,7 @@
         default: this.types=value?.split(','); break;
       }
       this.logOption('Timeline notifications', value);
-      if (!parent.userLoggedIn()) {
+      if (!parent.getUserId()) {
         parent.global.console.debug(' ... Disabled (not logged in)');
         return;
       }